refactor(web): extract the transcript fetch machine into useSessionTranscript
Phase 1 of #1327 (bound the eager transcript load with on-demand paging).
Pure move-out, no behaviour change: the initial history pull (single-session
backward-walk + merged-conversation fan-out), the live-tail refresh, and the
cross-source "load earlier" paging leave SessionDetailView for a dedicated
packages/web/src/lib/use-session-transcript.ts. The hook owns the transcript
state (msgs, loading/paging/error, tail readiness) and the conversation-mode
row bookkeeping; the view keeps deciding WHEN to refresh the tail and renders
the merged result. SessionDetailView sheds ~330 net lines.

Verified: tsc, eslint, prettier clean; session-transcript unit tests pass. The
happy-dom viewer suite can't run under this sandbox's Node 26 (localStorage
global vs happy-dom) — environmental, identical on base; it runs in CI on Node 24.
